Remove mm dependencies in `otx/core/utils` and move `mmconfig_dict_to_dict` into `otx/tools/translate_mmrecipe.py`
