+++ src/otx/tools/translate_mmrecipe.py
@@ -1,14 +1,19 @@
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2023-2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
 """Python script to translate MMx recipe to OTX YAML recipe file."""
+
+from __future__ import annotations
+
 from argparse import ArgumentParser
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 from mmengine.config import Config
 from omegaconf import OmegaConf
 
-from otx.core.utils.config import mmconfig_dict_to_dict
+if TYPE_CHECKING:
+    from mmengine.config import Config as MMConfig
 
 parser = ArgumentParser()
 parser.add_argument("-n", "--recipe-name", type=str, required=True)
@@ -19,6 +24,16 @@ override = parser.add_argument_group("override")
 override.add_argument("--base", type=str, default="detection")
 override.add_argument("--data", type=str, default="mmdet")
 override.add_argument("--model", type=str, default="mmdet")
+
+
+def mmconfig_dict_to_dict(obj: MMConfig | list[MMConfig]) -> list | dict:
+    """Convert MMEngine config object to Python dictionary."""
+    if isinstance(obj, list):
+        return [mmconfig_dict_to_dict(x) for x in obj]
+    if hasattr(obj, "to_dict"):
+        return {k: mmconfig_dict_to_dict(v) for k, v in obj.to_dict().items()}
+
+    return obj
 
 
 if __name__ == "__main__":
